Auto-detect suitable default behaviour for prefix canonicalization.
Current gcc offers -no-canonical-prefixes to turn off realpath() for prefixes
generated from the path used to address the gcc driver.  This allows gcc to
work in "symlink farm" installations, where every file in gcc is actually a
symlink to its real contents.  However, the flag has to be given explicitly.
If not, the default is to use realpath() to create prefixes and the result
is usually failure to find cc1[plus], f951, etc.

This patch adds a check for a file as a way to auto-detect whether prefix
canonicalization is appropriate or not.  Detection can be overridden by
using the -[no-]canonical-prefixes flags.

The patch also completes the fix for PR/29931, adding code that covers the
unadopted portion of this PR's attached patch.

gcc/ChangeLog.google:
2011-02-07  Simon Baldwin  <simonb@google.com>

	PR driver/29931
	* doc/invoke.texi: Adjust -[no-]canonical-prefixes documentation.
	* gcc.c (display_help): Help text for -[no-]canonical-prefixes.
	(driver_handle_option): Ignore OPT_canonical_prefixes.
	(process_command): Handle OPT_[no_]canonical_prefixes, auto-detect
	suitable default prefix canonicalization mode.
	* common.opt (canonical-prefixes): New flag.

Google ref: 40029, 38719



git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/google@169891 138bc75d-0d04-0410-961f-82ee72b054a4