This patch is for the google/gcc-4_7 and google/gcc-4_8 branches.
It fixes a problem with -fdebug-types-section and with
-gsplit-dwarf where the hash computation for the type signature
and for the DW_AT_dwo_id attribute may produce results that differ
from run to run. For dw_val_class_vec values (e.g., the const value
of a small struct), both hash computations use the value of a pointer
instead of the actual data.

Backported from trunk.

Google ref b/10289060

-cary


2013-08-13  Cary Coutant  <ccoutant@google.com>

gcc/
	* dwarf2out.c (CHECKSUM_BLOCK): New macro.
	(attr_checksum): Hash vector contents instead of pointer.
	(attr_checksum_ordered): Likewise.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/google@201714 138bc75d-0d04-0410-961f-82ee72b054a4