Remove DW_AT_GNU_addr_base from skeleton type unit DIEs.
GCC currently adds a DW_AT_GNU_addr_base attribute to skeleton type unit
DIEs, even though it's not needed there. By removing it, we can save
the 8 bytes that a DW_FORM_addr takes up, plus the 24 bytes that the
corresponding relocation takes.

This patch is for the google/gcc-4_8 branch. I will submit it
for trunk when stage 1 is open.

Tested with crosstool_validate.py.


2014-02-11  Cary Coutant  <ccoutant@google.com>

	* dwarf2out.c (add_top_level_skeleton_die_attrs): Don't add
        DW_AT_GNU_addr_base to all skeleton DIEs.
	(dwarf2out_finish): Add it only to the skeleton comp_unit DIE.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/google@207703 138bc75d-0d04-0410-961f-82ee72b054a4