Fix ICE when -gsplit-dwarf is used with -freorder-blocks-and-partition.
When FDO and -freorder-blocks-and-partition are enabled, it's possible
that by the time we get to optimize_location_lists, we have not yet
created any .debug_addr table entries. If -gsplit-dwarf is also enabled,
we still need to assign .debug_addr indexes to the location lists, so
we should be calling index_location_lists even if addr_index_table
is still NULL.

This patch is for the google/gcc-4_8 branch. I will checkin the fix
to trunk and the gcc-4_9 branch once I have a test case to exercise the
failure.

Google ref: b/15417905

gcc/
	* dwarf2out.c (dwarf2out_finish): Call index_location_lists
        even if addr_index_table is NULL.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/google@211247 138bc75d-0d04-0410-961f-82ee72b054a4