Mark std::unique, std::remove_if and std::remove with __attribute__((warn_unused_result)).
The attributes are added in the bits/algorithmfwd.h, which is included by the 
main implementation file bits/stl_algo.h.

While it's possible in peculiar circumstances to use these functions correctly 
without checking their return values, it's not something that's likely to
happen in practice.

(This also caught a bug in Clang code upstream, fixed separately in r234221.)

Google refs b/18886543, b/18980409.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/google@222073 138bc75d-0d04-0410-961f-82ee72b054a4